Django/anfisa1sprint/requirements.txt
asgiref==3.10.0
Django==3.2.16
django-debug-toolbar==4.2.0
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2025.2
sqlparse==0.5.3
